ELSA1-551 justerer oversikter i Storybook
